When casting between pointers and integers use intermediate ptrdiff_t.
On some 64-bit systems (notably Windows) long is still 32-bit so the
compiler could warn about different operand widths.

+++ src/client.c
@@ -920,7 +920,8 @@ store_result(struct command_state *state, int res)
 {
   int index = get_index(state);
   next_index(state);
-  state->object->store(state->object->arg, (void *) (long) res, index, NULL);
+  state->object->store(state->object->arg, (void *) (ptrdiff_t) res,
+                       index, NULL);
 }
 
 
@@ -1422,7 +1423,7 @@ state_prepare(struct command_state *state)
 
       while (count > 0)
         {
-          iov->iov_base = (void *) (buf + (long) (iov->iov_base));
+          iov->iov_base = (void *) (buf + (ptrdiff_t) (iov->iov_base));
           iov += step;
           count -= step;
         }
@@ -1847,7 +1848,7 @@ client_prepare_set(struct client *c, enum set_cmd_e cmd, int key_index,
     size_t str_size =
       sprintf(buf, " " FMT_FLAGS " " FMT_EXPTIME " " FMT_VALUE_SIZE "%s\r\n",
               flags, exptime, value_size, get_noreply(state));
-    iov_push(state, (void *) (long) array_size(c->str_buf), str_size);
+    iov_push(state, (void *) (ptrdiff_t) array_size(c->str_buf), str_size);
     array_append(c->str_buf, str_size);
   }
 
@@ -1888,7 +1889,7 @@ client_prepare_cas(struct client *c, int key_index,
       sprintf(buf, " " FMT_FLAGS " " FMT_EXPTIME " " FMT_VALUE_SIZE
               " " FMT_CAS "%s\r\n", flags, exptime, value_size, cas,
               get_noreply(state));
-    iov_push(state, (void *) (long) array_size(c->str_buf), str_size);
+    iov_push(state, (void *) (ptrdiff_t) array_size(c->str_buf), str_size);
     array_append(c->str_buf, str_size);
   }
 
@@ -1979,7 +1980,7 @@ client_prepare_incr(struct client *c, enum arith_cmd_e cmd, int key_index,
     char *buf = array_end(c->str_buf, char);
     size_t str_size =
       sprintf(buf, " " FMT_ARITH "%s\r\n", arg, get_noreply(state));
-    iov_push(state, (void *) (long) array_size(c->str_buf), str_size);
+    iov_push(state, (void *) (ptrdiff_t) array_size(c->str_buf), str_size);
     array_append(c->str_buf, str_size);
   }
 
@@ -2010,7 +2011,7 @@ client_prepare_delete(struct client *c, int key_index,
   {
     char *buf = array_end(c->str_buf, char);
     size_t str_size = sprintf(buf, "%s\r\n", get_noreply(state));
-    iov_push(state, (void *) (long) array_size(c->str_buf), str_size);
+    iov_push(state, (void *) (ptrdiff_t) array_size(c->str_buf), str_size);
     array_append(c->str_buf, str_size);
   }
 
@@ -2057,7 +2058,7 @@ client_flush_all(struct client *c, delay_type delay,
         size_t str_size =
           sprintf(buf, "flush_all " FMT_DELAY "%s\r\n",
                   (delay_type) (ddelay + 0.5), get_noreply(state));
-        iov_push(state, (void *) (long) array_size(c->str_buf), str_size);
+        iov_push(state, (void *) (ptrdiff_t) array_size(c->str_buf), str_size);
         array_append(c->str_buf, str_size);
       }
     }
